Python tends to convert my integers into floats in the generated code.
Former-commit-id: b0305e4362e3f150df278724b35e6547256f6732

+++ python/peano4/toolbox/blockstructured/ProjectPatchOntoFaces.py
@@ -35,7 +35,7 @@ class ProjectPatchOntoFaces(ActionSet):
       
     self.d[ "UNKNOWNS" ]           = str(patch.no_of_unknowns)
     self.d[ "DOFS_PER_AXIS" ]      = str(patch.dim[0])
-    self.d[ "OVERLAP" ]            = str(patch_overlap.dim[0]/2)
+    self.d[ "OVERLAP" ]            = str(int(patch_overlap.dim[0]/2))
     self.d[ "FACES_ACCESSOR" ]     = "fineGridFaces"  + patch_overlap.name
     self.d[ "CELL_ACCESSOR" ]      = "fineGridCell" + patch.name
     self.d[ "GUARD" ]                                        = guard
